requirements.txt
azure-storage-blob 
azure-cosmos 
openai 
python-dotenv
pypdf